+++ assets/js/_main.js
@@ -6,7 +6,7 @@
  * replace the dash with an underscore when adding it to the object below.
  *
  * .noConflict()
- * The routing is enclosed within an anonymous function so that you can 
+ * The routing is enclosed within an anonymous function so that you can
  * always reference jQuery with $, even when in .noConflict() mode.
  *
  * Google CDN, Latest jQuery
@@ -16,7 +16,7 @@
 
 (function($) {
 
-// Use this variable to set up the common and page specific functions. If you 
+// Use this variable to set up the common and page specific functions. If you
 // rename this variable, you will also need to rename the namespace below.
 var Roots = {
   // All pages
@@ -26,12 +26,6 @@ var Roots = {
 
       // Run fitVids on main content
       $('.main').fitVids();
-
-      // Wrap gform footer in Bootstrap ".form-group"
-      // TODO: Do this in php
-      $('.gform_footer').each(function() {
-        $(this).wrap('<div class="form-group"></div>');
-      });
 
     },
     finalize: function() { }