+++ lib/cleanup.php
@@ -259,3 +259,16 @@ function roots_get_search_form($form) {
   return $form;
 }
 add_filter('get_search_form', 'roots_get_search_form');
+
+/**
+ * Set low priorty for WordPress SEO metabox
+ */
+add_filter( 'wpseo_metabox_prio', function() {
+  return 'low';
+});
+
+
+/**
+ * Remove WordPress SEO columns
+ */
+add_filter( 'wpseo_use_page_analysis', '__return_false' );