+++ lib/gallery.php
@@ -40,7 +40,7 @@ function roots_gallery($attr) {
     'itemtag'    => '',
     'icontag'    => '',
     'captiontag' => '',
-    'columns'    => 4,
+    'columns'    => 3,
     'size'       => 'thumbnail',
     'include'    => '',
     'exclude'    => '',
@@ -48,7 +48,7 @@ function roots_gallery($attr) {
   ), $attr));
 
   $id = intval($id);
-  $columns = (12 % $columns == 0) ? $columns: 4;
+  $columns = (12 % $columns == 0) ? $columns : 3;
   $grid = sprintf('col-sm-%1$s col-lg-%1$s', 12/$columns);
 
   if ($order === 'RAND') {
@@ -100,7 +100,7 @@ function roots_gallery($attr) {
     $output .= '<div class="' . $grid .'">' . $image;
 
     if (trim($attachment->post_excerpt)) {
-      $output .= '<div class="caption hidden">' . wptexturize($attachment->post_excerpt) . '</div>';
+      $output .= '<div class="caption">' . wptexturize($attachment->post_excerpt) . '</div>';
     }
 
     $output .= '</div>';
@@ -124,7 +124,7 @@ if (current_theme_supports('bootstrap-gallery')) {
  */
 function roots_attachment_link_class($html) {
   $postid = get_the_ID();
-  $html = str_replace('<a', '<a class="thumbnail img-thumbnail"', $html);
+  $html = str_replace('<a', '<a class="thumbnail"', $html);
   return $html;
 }
 add_filter('wp_get_attachment_link', 'roots_attachment_link_class', 10, 1);